data_processor/pyproject.toml
[project]
name = "data_processor"
version = "0.1.0"
description = "A CouchDB change-feed-driven data processing service"
authors = [{ name = "Rob Sanders", email = "developer@zendo.com" }]
requires-python = "~=3.12.0"
dependencies = [
    "aiohttp>=3.12.13",
    "pandas>=2.3.0",
    "python-dotenv>=1.1.0",
]

[dependency-groups]
dev = [
    "ruff>=0.12.0",
]
